Gl doom sdl2

+++ m_fixed.c
@@ -42,7 +42,7 @@ rcsid[] = "$Id: m_bbox.c,v 1.1 1997/02/03 22:45:10 b1 Exp $";
 
 fixed_t FixedMul( fixed_t a, fixed_t b )
    {
-    return (((__int64) a * (__int64) b) >> FRACBITS);
+    return (((long long) a * (long long) b) >> FRACBITS);
    }
